requirements.txt
mailru-im-bot
python-dateutil
pendulum
pyaspeller
tarantool
dateparser